refactor Bank to EthereumApp and ERC20App
