+++ ethereum/contracts/EthereumApp.sol
@@ -2,19 +2,15 @@
 pragma solidity >=0.6.2;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
-contract Bank {
+contract EthereumApp {
     using SafeMath for uint256;
 
     uint256 public nonce;
     uint256 public totalETH;
-    mapping(address => uint256) public totalTokens;
 
-
-    enum AppEventTags { SendETH, SendERC20 }
-
-    event AppEvent(uint _tag, bytes _data);
+    event AppEvent(bytes _data);
+    event Unlock(address _recipient, uint256 _amount);
 
     constructor() public {
         nonce = 0;
@@ -33,28 +29,7 @@ contract Bank {
         nonce = nonce.add(1);
 
         bytes memory data = encodeSendData(msg.sender, _recipient, address(0), msg.value, nonce);
-        emit AppEvent(uint(AppEventTags.SendETH), data);
-    }
-
-    function sendERC20(
-        bytes32 _recipient,
-        address _tokenAddr,
-        uint256 _amount
-    )
-        public
-    {
-       require(
-            IERC20(_tokenAddr).transferFrom(msg.sender, address(this), _amount),
-            "Contract token allowances insufficient to complete this lock request"
-        );
-
-        // Increment locked ERC20 token counter by this amount
-        totalTokens[_tokenAddr] = totalTokens[_tokenAddr].add(_amount);
-        // Increment global nonce
-        nonce = nonce.add(1);
-
-        bytes memory data = encodeSendData(msg.sender, _recipient, _tokenAddr,_amount, nonce);
-        emit AppEvent(uint(AppEventTags.SendERC20), data);
+        emit AppEvent(data);
     }
 
     function encodeSendData(
@@ -69,5 +44,30 @@ contract Bank {
         returns(bytes memory)
     {
         return abi.encode(_sender, _recipient, _tokenAddr, _amount, _nonce);
+    }
+
+     function submit(bytes memory data)
+        public
+    {
+        // TODO: decode message bytes into (tag, recipient, amount)
+         uint256 tag = 0;                        // placeholder
+         address payable recipient = address(0); // placeholder
+         uint256 amount = 1;                     // placeholder
+
+        unlockETH(recipient, amount);
+    }
+
+    function unlockETH(
+        address payable _recipient,
+        uint256 _amount
+    )
+        internal
+    {
+        require(_amount > 0, "Must unlock a positive amount");
+        require(totalETH > _amount, "ETH token balances insufficient to fulfill the unlock request");
+
+        totalETH = totalETH.sub(_amount);
+        _recipient.transfer(_amount);
+        emit Unlock(_recipient, _amount);
     }
 }
